Hibernate example and AccountServiceImpl added
